Support for subcommands, renamed pet (For users) to minion. added info
I added subcommand support so that /petname and /minionname are both able to be used (I will 100% personally only use /petname)
Renamed all instances of pet to minion for the users.
Added a little info spot that informs users to look away or resummon the pet to apply the name changes.